refactor(vue): from ts to jsdoc
