+++ packages/vue/src/app/index.d.ts
@@ -1,0 +1,1 @@
+export declare const AppProvider: import("react").ForwardRefExoticComponent<any>;
