+++ packages/vue/src/app/index.js
@@ -1,0 +1,7 @@
+import { onUnmounted, $bridge } from '../index';
+export const AppProvider = $bridge((props) => {
+    onUnmounted(() => {
+        props.app.unmount();
+    });
+    return () => props.children;
+}, "AppProvider");
