--- packages/vue/src/constants.ts
@@ -1,6 +1,0 @@
-
-export const IS_GLOBAL_KEY = Symbol('__v_isGlobal');
-export const INITIAL_CONTEXT_KEY = Symbol('__v_initialContext');
-export const CONTEXT_KEY = Symbol('__v_context');
-export const SUBSCRIBE_KEY = Symbol('__v_subscribe');
-export const LISTENER_LIST_KEY = Symbol('__v_listenerList');
