+++ packages/vue/src/hook-manager/index.js
@@ -1,6 +1,6 @@
 import { HookManager, usePlugin } from "@bridge/core";
 import HookRef from "../react-hook/hookRef";
-import { unref } from "@vue-internals/reactivity/ref";
+import { unref } from "#vue-internals/reactivity/ref";
 
 usePlugin(HookManager, { signalClass: HookRef, unsignal: unref });
